Assert all nodes are entrypoints
Not all rbx ast nodes are root nodes of an syntax tree. For example
there is no source that would result in an instance of
Rubinius::AST::FormalArguments19 beeing a root node.

Mutant and other tools might generate such code for internal use. So we
test all reachable nodes of an ast to produce round trippable source.